Few minor fixes.

+++ include/serialfmm.h
@@ -22,7 +22,7 @@ private:
 #endif
     real_t x = 1.0 / THETA;                                     // Inverse of theta
 #if ERROR_OPT
-    real_t a = c * pow(std::abs(C->M[0]),1.0/3);                // Cell coefficient
+    real_t a = c * powf(std::abs(C->M[0]),1.0/3);               // Cell coefficient
     for (int i=0; i<5; i++) {                                   // Newton-Rhapson iteration
       real_t f = x * x - 2 * x + 1 - a * pow(x,-P);             //  Function value
       real_t df = (P + 2) * x - 2 * (P + 1) + P / x;            //  Function derivative value
@@ -121,7 +121,7 @@ public:
     Ci0 = cells.begin();                                        // Set iterator of target root cell
     Cj0 = cells.begin();                                        // Set iterator of source root cell
     upwardRecursion(Ci0, Ci0);                                  // Recursive call for upward pass
-    real_t c = (1 - THETA) * (1 - THETA) / pow(THETA,P+2) / pow(std::abs(Ci0->M[0]),1.0/3); // Root coefficient
+    real_t c = (1 - THETA) * (1 - THETA) / pow(THETA,P+2) / powf(std::abs(Ci0->M[0]),1.0/3); // Root coefficient
     setRcrit(Ci0, Ci0, c);                                      // Error optimization of Rcrit
     for (C_iter C=cells.begin(); C!=cells.begin()+9; C++) {     // Loop over top 2 levels of cells
       C->RCRIT *= 10;                                           //  Prevent approximation
